Move compiler class to new file Rewrite try ... catch operators - now program will not fall with exceptions i run-time but will be printed error message and program will be stopped
